Define warning() when in debug mode
I noticed that multiple places in the code did had the warning function
commented out, so instead of just removing them the better solution
was to use the already existing release/debug system. When compiling
under debug mode, CMake defines DEBUG=1 which in turn enables the
warning() function. Otherwise it's an empty macro.